Leves cambios. O mais importante é que se engadiu unha entrada de modo de emprego ao arquivo README

+++ stop_services.bash
@@ -1,14 +1,17 @@
 #! /bin/bash
 
+if [ $# -lt 2 ]
+then
+    echo "The number of arguments is invalid. You need to specify two Interface names, one for AP and other connected to internet. In that order."
+    exit
+fi
+
 systemctl stop isc-dhcp-server
 systemctl stop hostapd.service
 
-#iptables -t nat -D POSTROUTING -o enp1s0 -j MASQUERADE
-#iptables -D FORWARD -i wlx002719b8dec1 -o enp1s0 -j ACCEPT
+iptables -t nat -D POSTROUTING -o $2 -j MASQUERADE
+iptables -D FORWARD -i $1 -o $2 -j ACCEPT
 
-iptables -t nat -D POSTROUTING -o wlp0s20f3 -j MASQUERADE
-iptables -D FORWARD -i wlx002719b8dec1 -o wlp0s20f3 -j ACCEPT
+ip address delete 192.168.101.1/24 dev $1
 
-ip address delete 192.168.10.1/24 dev wlx002719b8dec1
-
-ip link set wlx002719b8dec1 down
+ip link set $1 down
